+++ styles/home.tsx
@@ -1,0 +1,32 @@
+import { StyleSheet } from 'react-native'
+
+export default StyleSheet.create({
+    view: {
+        justifyContent: 'center',
+        alignItems: 'center',
+    },
+    logo: {
+        width: '90%',
+        height: '20%',
+    },
+    backgroundImage: {
+        width: '112%',
+        height: '112%',
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    buttons: {
+        width: 200,
+        height: 50,
+        backgroundColor: '#11263B',
+        justifyContent: 'center',
+        padding: '2%',
+        borderRadius: 5,
+    },
+    buttonsText: {
+        color: 'white',
+        fontSize: 13,
+        fontFamily: 'Poppins_400Regular',
+        textAlign: 'center',
+    },
+})
