+++ types.d.ts
@@ -1,8 +1,14 @@
 type CreateUserRequestBody = {
     device_id: string
     should_send_notifications?: boolean
-    time_trigger: number
+    time_trigger?: number
     expo_push_token: string
+}
+
+type UpdateUserRequestBody = {
+    should_send_notifications?: boolean
+    time_trigger?: number
+    expo_push_token?: string
 }
 
 type SaintInfo = {
@@ -25,4 +31,16 @@ type Phrase = {
 type TimeTrigger = {
     hour: number
     minute: number
+}
+
+type User = {
+    device_id: string
+    should_send_notifications: boolean
+    time_trigger: number
+    expo_push_token: string
+}
+
+type GlobalContextProps = {
+    user: User
+    setUser: React.Dispatch<React.SetStateAction<User>>
 }
