+++ utils/consts.ts
@@ -1,8 +1,3 @@
-export const defaultTrigger: TimeTrigger = {
-    hour: 8,
-    minute: 0,
-}
-
 export const blue = '#024959'
 export const lightblue = '#4a868c'
 export const crossblue = '#11263B'
