+++ utils/users.ts
@@ -6,14 +6,21 @@ const getUser = async (deviceId: string) => {
     return axios
         .get(API_URL + '/user/' + deviceId)
         .then(res => res.data)
-        .catch(err => console.error(err))
+        .catch(err => console.error(err.message))
 }
 
 const createUser = async (body: CreateUserRequestBody) => {
     return axios
         .post(API_URL + '/user/' + body.device_id, body)
         .then(res => res.data)
-        .catch(err => console.log(err))
+        .catch(err => console.error(err.message))
 }
 
-export { getUser, createUser }
+const updateUser = async (deviceId: string, body: UpdateUserRequestBody) => {
+    return axios
+        .put(API_URL + '/user/' + deviceId, body)
+        .then(res => res.data)
+        .catch(err => console.error(err.message))
+}
+
+export { getUser, createUser, updateUser }
